+++ CHANGELOG.md
@@ -15,6 +15,7 @@ v1.6.1 [unreleased]
 -   [#257](https://github.com/latera/homs/pull/257) Check flag `multi` in `select_table` when select row.
 -   [#262](https://github.com/latera/homs/pull/262) Fix multi nullable `select_table`.
 -   [#265](https://github.com/latera/homs/pull/265) Allow nullability for `select` lookup mode.
+-   [#264](https://github.com/latera/homs/pull/264) Fix required `select` with numeric value.
 
 v1.6.0 [2019-01-29]
 -------------------